move CapturingLogHandler class from subset to loggingTools module
