+++ Lib/fontTools/subset/subset_test.py
@@ -2,7 +2,7 @@ from __future__ import print_function, division, absolute_import
 from fontTools.misc.py23 import *
 from fontTools import subset
 from fontTools.ttLib import TTFont, newTable
-import contextlib
+from fontTools.misc.loggingTools import CapturingLogHandler
 import difflib
 import logging
 import os
@@ -10,38 +10,6 @@ import shutil
 import sys
 import tempfile
 import unittest
-
-
-class CapturingLogHandler(logging.Handler):
-  def __init__(self, log_path, level):
-    self.records = []
-    self.level = level
-    self.logger = logging.getLogger(log_path)
-
-  def __enter__(self):
-    self.original_disabled = self.logger.disabled
-    self.original_level = self.logger.level
-
-    self.logger.addHandler(self)
-    self.logger.level = self.level
-    self.logger.disabled = False
-
-    return self
-
-  def __exit__(self, type, value, traceback):
-    self.logger.removeHandler(self)
-    self.logger.level = self.original_level
-    self.logger.disabled = self.logger.disabled
-    return self
-
-  def handle(self, record):
-    self.records.append(record)
-
-  def emit(self, record):
-    pass
-
-  def createLock(self):
-    self.lock = None
 
 
 class SubsetTest(unittest.TestCase):
